+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.0.4] - 2019-05-20
+### Changed
+- Comparison of the onAllReady function prop was causing an infinite loop.
+For now, ignore that property when checking for updates. This is a hack.
+
 ## [0.0.3] - 2019-05-20
 ### Added
 - onAllReady callback